Change CallActivity to use an abstraction to iterate calls
